Remove unnecessary prefixes/suffixes from column names (#10)
- Adds a generalizable approach to safely renaming columns to remove
unwanted prefixes and suffixes.
- The `SUBSTRINGS_TO_FIX` are configured as a list in the constants
module
- Added `build_substring_removal_clauses()` which is called from
`process_columns()` to handle these removals
- Ensures that the renamed column does not collide with another column
with the same name unless it is appropriately consolidated with a
coalesce statement.
- Refactored process columns so that intermediate tables are not stored
unnecessarily. Instead, the affected variable names and the select
statements are passed to the main function

Minor changes: 
- renamed functions called from process_columns to better reflect there
function
- added `utils.standardize_column_case()` to ensure that all column
names (except "Connect_ID") are all lowercase in the resulting tables

Closes
[#24](https://github.com/Analyticsphere/pr2-documentation/issues/24)