Functionality to identify and correct "false array" variables (#14)
- Added functionality to `process_rows()` to remove brackets from "false
array" variables
- `reference/false_array_columns.json` contains a list of concept_ids
that identify columns that are known to contain "false arrays"
- `utils.load_false_array_reference()` loads this file 
- `utils.get_false_array_columns_from_reference()` identifies false
array columns by checking if they contain concept ID pairs from the
reference file
- `utils.get_strict_false_array_columns()` can retrieve a list of false
array columns from a table using the reference file or optionally using
an algorithm. Using the reference file is faster.
- `utils.render_unwrap_singleton_expression()` renders a SQL expression
to handle survey unwrapping singleton values from variables stored as
"false arrays".
- `scripts/idenitify_false_array_columns.py` can be used to generate a
new reference file containing the concept ids that are used to identify
false array columns. It needs a little work (in a seperate PR) but is
unlikely to be used again as the underlying issue appears to be
resolved.

Closes
[#27](https://github.com/Analyticsphere/pr2-documentation/issues/27)